[FIX] Movie function & Movie Page
