+++ src/Pages/Movies/Movie_Components/CrewCard.tsx
@@ -1,6 +1,7 @@
 import { Card, CardContent, CardMedia, Typography, CardActions, Button, Box } from "@mui/material";
 import logo from '../../../assets/logo/TheWatcher.png';
 import { useNavigate } from "react-router-dom";
+import { FrJob } from "../../../assets/functions/crew_functions";
 
 function CrewCard (props: any) {
 
@@ -21,21 +22,6 @@ function CrewCard (props: any) {
             display += fr_job[i] + ', ';
         }
         return display.slice(0, -2);
-    }
-
-    const FrJob = (job: String) => {
-        switch (job) {
-            case 'Director':
-                return 'Réalisateur';
-            case 'Screenplay':
-                return 'Scénariste';
-            case 'Writer':
-                return 'Scénariste';
-            case 'Original Music Composer':
-                return 'Compositeur';
-            default:
-                return job;
-        }
     }
 
     return (
