+++ src/assets/functions/movie_functions.ts
@@ -9,7 +9,18 @@ const customDate = (date: String) => {
 }
 
 const getStatus = (status: String, release_date: String) : String => {
-    return (status === 'Released') ? 'Sorti le ' + customDate(release_date) : 'A Sortir le ' + release_date;
+    switch (status) {
+        case 'Rumored':
+            return 'Rumeur';
+        case 'Post Production':
+            return 'Post production';
+        case 'Released':
+            return 'Sorti le ' + customDate(release_date);
+        case 'Canceled':
+            return 'Annulé';
+        default:
+            return 'Inconnu';
+    }
 }
 
 export { getRating, customDate, getStatus}
